Solves issue after recovering from L4 data init not from local FS
Details:
During L4 recovery in FTI_Init, the ckpt files are moved to the local FS
The recovery of the data during FTI_Recover, however, was still done from
the PFS. That made the transfer to local during FTI_Init meaningless.

With this commit, the ckpt data is recovered from L1 directory during
FTI_Recover.