Fixed bug which caused hanging
If the snapshot function was called while kernels were being executed with the
backup mechanism this caused all processes to hang. This is because a call is
made to MPI_allreduce() at some point during a snapshot.  Since MPI_allreduce
is a collective operation it waits for all processes to make this call. The
problem with this is some processes may have finished executing their kernels
before others and can potentially never call snapshot() again and by extension,
never calling MPI_allreduce again.

The kernel backup mechanism works by executing a kernel in a loop so that
snapshots can be made each time it is interrupted. So to solve the problem of
some kernels finishing before others and causing a hang, no process is allowed
to exit the kernel execution loop until all kernels are complete.

This is achieved by using MPI_Allgather() and a boolean array of a size equal
to the number of MPI processes. When each kernel is complete it writes a true
value to the array. When the array has a number of true values equal to the
number of processes launched, it means all processes have finished executing
their kernels and the loop can now exit.