Removed global variables from fti.h for kernel interrupt
This cleans up the code a bit and no longer exposes global variables
for accidental access. The FTIT_gpuInfo struct is not an array of size FTI_BUFS
in FTIT_execution and kernels can now be protected with the macro
FTI_Protect_Kernel. The macro takes an ID which is useful for keeping track
of which kernel is being protected.