Runs but needs to be changed
The modifications made to CreateMetadata have a conceptual flaw, by
using MPI_Gather to collect the GPU metadata it is not possible to
determine the kernel ID for the collected data. Because of this, FTI_CreateMetadata
will need to be modified in a different way. This version runs correctly for now,
but has to be changed. Currently the metadata collected using MPI_Gather is not
written to file.