Added Gpu metadata structure
This structure is useful in capturing the metadata when suspending
kernels so that it can be written at checkpoint time.

Now using MPI_Send and MPI_Recv to capture GPU metadata instead of
MPI_Gather since MPI_Gather is a collective operation and does not
preserve the information necessary to partition and write the GPU
metadata.

Testing of this is still needed.