Calculating total number of application processes
To determine the total number of processes the nbProc member in
FTIT_topology was used. This was however discovered to be incorrect
if FTI is configured to run with an FTI dedicated (head) process
per node. The number of application processes is now being determined
by multiplying the nbApprocs and nbNodes members in FTIT_topology.