Relaunching completed kernel only if all others are complete
By default a previously complete kernel will have its metadata reset
and will be relaunched. However, in the case of a restart this can
cause a kernel to unnecessarily re-execute. This modification checks
all registered kernels so far for completion. Only if all kernels are
complete a kernel is allowed to be relaunched; this ensures that execution
continues from the last incomplete executing kernel.