Added limit to how much quantum of protected kernel is automatically increased.
The quantum is increased if the kernel makes no progress between interruptions.
This is determined by keeping track of how many blocks were executed between
the last and current interruption. If the number of blocks does not increase then
the quantum, if below the limit, is increased.

The reason for this is because a very small quantum expire before the kernel has
had the chance to schedule an previously unexecuted block. However, the quantum
cannot be allowed to continiously increase as this risks the chance of greatly
lengthening the kernel quantum because there may be other reasons (such as multiple
kernels executing on a device) that causes the kernel not to progress.

The quantum is set to a limit of 5 seconds which is more than enough time for a
kernel to launch all of its blocks.