Now using array of pointers to KernelInfo instead of array of KernelInfo
This change was made within FTI_execution since it reduces the number
of lines of code to keep track of and restore kernel information.

Without pointers, the fields of the FTI_kernelInfo structure have to be tracked
member by member. By using pointers the address of the structure can be taken instead.
This has the advantage of reducing the number of lines of code to maintain while making
it straightforward to add new members to the FTI_KernelInfo struct if necessary.