Counting allocations of non-restarted protected kernels.
Allocations for protected kernels are done each time a new kernel
is added for protection. Whereas allocations for kernels to be restored
are done in bulk up to the maximum number of kernels that can be protected.
These differing allocations schemes also require differing freeing schemes.
There needs to be a call to free for each protected kernel that has not been
restored from a checkpoint. Simarily there needs to be a single call to free
for all kernes restored from a checkpoint. The latter was already happening, so
this commit counts the number of allocations of non-checkpointed kernels so that
if the application runs to completion without needing a restart, the memory allocated
per protected kernel can be freed.