Removed argument from function that adds kernel metadata.
The index of the current group was being passed to be used as an index when
adding kernel's metadata to dictionary. However, it was realised that the index
was being for another argument being passed to the FTI_AddKernelMetadata also.
So this could be shorted by just performing the indexing right when
FTI_AddKernelMetadata is being called.