Tutorial overhaul and small bugfixes in example app
The tutorial have been overhauled and contains the following changes:

1) Readability and Text Standarization

The tutorial is now written in a standard format with the following characteristics.
One sentence per line, mostly in imperative first person with all scripts in inline or multiline blocks.

2) Installation script and .gitignore changes

A script for installing FTI is provided for easy installation for tutorial purposes in scripts/install.sh
The script is extracted from the previous tutorial text and does not depend on environment variables.
The variables are still supported to maintain portability.
In the script, the build and install directories default to the FTI root.
As such, these folders are ignored in .gitignore so that contributors do not accidentally submit files generated by cmake.
Also, the Makefiles under test/ and run scripts generated by cmake are also added to .gitignore

3) Questions and Answers

Questions in the tutorial are now answered and put inside a Spoiler with quotes.
A reference to the paper is mentioned in the end of the Practice section for more information.

4) Bugfix in the HeatDist application in the examples

A bug in 32 bits machines is corrected in the heatdist application in the examples.
The bug is related to wrongly calculating the size of the datastructures and was not impacting x86 systems.
This was due to double and pointer having the same size in these systems whereas pointers have 32 bits in 32 machines.

5) Bugfix in the Tutorial Makefiles for low core count machines

A bug is corrected in the tutorial for machines with less than 8 cores.
OpenMPI does not oversubscribe and does not account for hardware threads by default.
As such, testing FTI with less than 8 cores involved changing the makefiles manually.
Now, the settings are the same but the --oversubscribe parameter is passed to mpirun.
This way, the examples will run on any machine at the cost of performance (should be irrelevant for testing purposes).

6) Bugfix in Syntax of Bash commands in Tutorial

Changed the watch command in the L1 example.
It had a wrong syntax and are now working with bash syntax.
Also, now it is explicited that another Terminal will be needed to watch changes as the application runs.