Add IME native API I/O calls for level 4 checkpointing
With this patch we add an additional I/O interface that can be use to
perform level 4 checkpoints. The new API added is the IME native I/O
interface that is used to link the applications I/O directly with the
IME client library bypassing fuse.

DDN's Infinite Memory Engine® (IME®) is a scale-out, software-defined,
flash storage platform that streamlines the data path for application I/O.