Added ITF filters to ignore tests by parameter or function name
- Tests can be be disabled by name in the testdriver using --ignore/-i 'testname';
- Test cases can be disabled by parameter value in the testdriver using --filter/-f 'suite:test:parameter=value';
- RecoverName tests are disabled in CI when IO is FTI-FF using this new feature;
- Standard tests are disabled in CI when expected outcome of the test is 'failure';

The tests that are disabled in CI still make part of the suite and will be launched when doing local tests.
This new feature allows for more flexibility in regards to which tests we want to run.
It does not alter the test suites nor the engine.
Plus, scripts embedded with filters are more verbose in terms of what it is actually doing in comparison to the old suite files.