Bugfix on failing compilation when OpenSSL is not found
Prior to this commit, compilation would fail when openssl was not found in the system.
The reason lays in the fact that test applications had some logic to include either openssl or FTI's md5 module.
This commit reworks how md5 is integrated into the project by simplifying CMake variables and include statements.

The idea is that md5.h will always be included, regardless of openssl being found or not.
The header will either include openssl, or define the custom FTI md5 implementation as necessary.
To simplify compilation, the md5.h.in file is removed as configuration can be entirely using define statements.
To correctly issue these statements, CMake has been modifyied to use only one new variable: HAVE_OPENSSL.
All the changes are in place to make this new interaction work as before.

- Fixed test applications include statements;
- Reworked CMakeLists files for md5 and root folders;
- Removed variables with overlapping functionality regarding the inclusion of OpenSSL;
- Simplified the compilation for md5 module;
- Transformed md5 module into an indirection layer as previously intended;