Bundle of changes to address new docker image with code coverage support
- CMake has been expanded with a new ENABLE_COVERAGE option;

This option links gcov and add compile flags to all FTI targets.
ENABLE_COVERAGE is disabled by default and only works when the compiler is GCC.
The new flags compute which lines of code have been executed in FTI.
Each application linked with FTI will contribute to these counters, making it perfect to use with current test suites.

- Added a new option, --enable-coverage, for the install script;

This option, disabled by default, sets the ENABLE_COVERAGE option during build.

- Added a new option, --make-verbose, for the install script --make-verbose;

This option sets the verbosity of the make command to true.
The best use for this option is to debug the build and installation procedure for FTI.

- Added the --enable-coverage option into the CI build script for the GCC compiler;

This change will make the CI environment be able to generate coverage metrics while executing the checks.

- Created a new testing tool program to generate xml coverage metrics;

- Changed CMakeList target_link_library usage to match CMake documentation;

The use of target_link_library does not include specifying multiple targets.
As such, it should be called once per target and not how it was being called.
The consequence was that FTI was being marked a dependency of FTI.
Although this was not proeminent on older versions of gcc, this is a problem on gcc-10.

- Adjusted the compilation test suite to test CMake versions 3.3.2 through 3.16;

- Re-introduced the possibility to include empty entries to an ITF string list;

This is necessary to keep track of test cases that have no arguments (i.e sync_intv).

- Bugfix on ITF functions for corrupting checkpoints for HDF5 IO;

The function now crop the file extension before using the rank number to match the files.
This process would fail on hdf5 checkpoints because HDF5 files contain numbers on the file extension (i.e .h5).