Removed nameref bash variables from ITF source code
This functionality is available in bash v4.3.3 and onwards.
However, as they are not critical, we decided to remove them to lower the compatibility.
Nameref variables work as indirect pointers to bash variables.
They are heavily used on ITF custom list implementation on the following methods.

- itf_list_unwrap;
- itf_list_add;
- itf_list_contains;
- itf_list_remove;
- itf_array_contains.

This commit changes the nameref variables to bash v2.0 indirect ${!references} and the POSIX 'eval' command.
The change will make the code less safe but will provide the goal of lowering requirements from bash 4.3 to bash 4.0.