Compilation suite bugfix and minor corrections when compiling with gcc-10.1
Changelog

- Installation script now returns non-zero when make or cmake fails as intended

The previous verion of the install script did not work as intended when instructions were piped.
In this scenario, it is necessary to use the BASH variable PIPESTATUS.

- Added external to the `granks' variable in diff_test.h

The old version failed to compile on gcc 10.1 as this rule is stricter in this version.

- Added the --cmake-bin option in the install script

This option enables compilation to happen with a custom cmake binary.
The ideia is that this command allows us to use, and test, this script in CI.
Also, It might help people using non-standard distributions of CMake.

- Compilation test suite now uses different CMake versions as intended

Previous test behavior did only check if the cmake binary existed.
When compiling, it used the system's default cmake installation.
Now the tests uses the new --cmake-bin option in the installation script.