Build and compilation fixes for gcc-10.1 and cmake-3.17
- Installation script now returns non-zero when make or cmake fails as intended;

The install script failed to return non-zero when operating on verbose mode.
The cause was piped bash commands which return their status via the PIPESTATUS variable and not `$?'.

- Fixed compilation error for gcc-10.1 in diff_test application;

The error is related to two files including the same header and defining the same variable.
CMake compiles files isolated from each other.
As such, header guards (i.e _DIFF_TEST_H_) are not effective and will end up with duplicate variable definitions.

- Added the --cmake-bin option in the install script;

This option enables compilation to happen with a custom cmake binary.
The ideia is that this command allows us to use, and test, this script in CI.
Also, It might help people using non-standard distributions of CMake.

- Compilation test suite now uses different CMake versions as intended

Previous test behavior did only check if the cmake binary existed.
When compiling, it used the system's default cmake installation.
Now the tests uses the new --cmake-bin option in the installation script.

- Removed append_property custom CMake function from CMakeLists.

This method was an attempt to bring compatibility to CMake version 2.8.
However, the minimum CMake version for FTI is now 3.3.
This patch replaces the functionality with CMake standard functions.