Added mappings from Fortran types to native C types when using HDF5 IO
The solution is based on two steps, the bash pre-processor (BPP) and a initialization C function in the interface.
Upon compilation, BPP lists the datatypes supported by the compiler along with their size in bytes.
Then, BPP generate calls to a new C function in the Fortran interface wrapper, the FTI_InitPrimitiveType.
This function parses the Fortran datatype name and associate them to C datatype names (e.g. real may map to float/double/long double and integer may map to short/int/long).
The function uses sizeof() operations to check which relevant C type has the same size as its Fortran counterpart.
If a match is possible, FTI re-uses the FTIT_Type associated with that primitive.
Otherwise, a new custom FTIT_Type is created and therefore a new non-native HDF5 datatype.

Warning
The solution hereby proposed is based on the premise that both C and Fortran share a common binary format for similar datatypes.
That is, floating point datatypes should be in the IEEE 754 format and integer/logical datatypes should be in the two's complement format.
This should be the case on most compilers and systems.
However, if there is no such correlation, this feature will fail silently and store HDF5 checkpoints in respect to C-like datatypes.
A sympton of this problem is that Fortran applications recovering from checkpoints will have its data corrupted.