Changed the datatype definition in FTI to use native C datatypes
The previous use of fixed-size datatypes could present harm to integration with IO libraries.
As an example, the FTI_INT type would have the size of the int32_t type, which is garanteed to be 32 bits.
In a system where integers are 64 bits, HDF5_NATIVE_INT will have the size of int64_t which would be the same as int.
However, because we mapped FTI_INT to int32_t, its size would be 4 bytes.
The integrations between FTI and HDF5 links the datatypes FTI_INT and HDF5_NATIVE_INT, which would be incoherent in regard to size.
Using the native C types, we can expect equivalence to other library native datatypes.
In other words, FTI_INT maps to 'int' which maps to HDF5_NATIVE_INT, in our example, regardless of the platform-specific size of 'int'.