Added a new test suite to check datatype mapping from Fortran to C
The test suite checks, for all IOs, if Fortran meta-dada is the same as C meta-data.
Two applications are developed with the same behavior, one in each language.
Both applications execute and we compare the meta-data files as they should be the same.
Then, the applications recover their data using FTI to guarantee that all primitve datatypes are working as intended.