Rework on datatype handling API and internal functions
The datatype handling routines, specially for HDF5, are now opaque to the user.
FTIT_type objects are no longer directly accessible to the user as intended.
Users will now register custom data types and receive a handle for that object (i.e fti_id_t, which is an int).
FTI will manage the allocation, initialization and destruction of these types in behalf of the user.
These changes affect the public API and many of the tests had to be rewritten to the new API.